# Backend/docs/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Volcanoes Around the World! v2.0",
        "description": "This Swagger API document exposes a number of REST endpoints which implement CRUD operations on a database containing publicly available data about volcanoes collated by the Smithsonian Institution's Global Volcanism Program. The API endpoints and their usage are described in detail below.",
        "contact": {
            "email": "n11078472@qut.edu.au"
        }
    },
...
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token only:"
            }
        }        
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
        "/countries": {
            "get": {
                "tags": [
                    "Data"
                ],
                "summary": "Retrieve country data",
                "description": "Returns a list of all countries that are associated with one or more volcanoes, ordered alphabetically.",
                "security": [],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "An array of countries. An example of the first five elements in the array is shown below.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    },
                                    "example": [
                                        "Algeria",
                                        "Antarctica",
                                        "Argentina",
                                        "Armenia",
                                        "Australia"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Query parameters not permitted."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server / database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error in MySQL query"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/volcanoes": {
            "get": {
                "tags": [
                    "Data"
                ],
                "summary": "Retrieve volcano data from a specific country",
                "description": "Returns a list of volcano objects, each with an id, name, country, region, and subregion. The country query parameter is required. Optionally, the list can be filtered using the populatedWithin query parameter, to reduce the list to volcanoes with at least one person living within the provided radius.",
                "security": [],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "country",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Name of the country to filter volcanoes."
                    },
                    {
                        "name": "populatedWithin",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "5km",
                                "10km",
                                "30km",
                                "100km"
                            ]
                        },
                        "description": "Optional parameter to filter volcanoes based on population proximity."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array of volcano objects, each containing an id, name, country, region, and subregion. An example array is shown below.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Abu"
                                            },
                                            "country": {
                                                "type": "string",
                                                "example": "Japan"
                                            },
                                            "region": {
                                                "type": "string",
                                                "example": "Japan, Taiwan, Marianas"
                                            },
                                            "subregion": {
                                                "type": "string",
                                                "example": "Honshu"
                                            }
                                        }
                                    }
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "name": "Abu",
                                        "country": "Japan",
                                        "region": "Japan, Taiwan, Marianas",
                                        "subregion": "Honshu"
                                    },
                                    {
                                        "id": 16,
                                        "name": "Aogashima",
                                        "country": "Japan",
                                        "region": "Japan, Taiwan, Marianas",
                                        "subregion": "Izu, Volcano, and Mariana Islands"
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters. Click on 'Examples' dropdown below to see possible error responses.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "MissingCountryParameter": {
                                        "summary": "MissingCountryParameter",
                                        "value": {
                                            "error": true,
                                            "message": "Country is a required query parameter"
                                        }
                                    },
                                    "InvalidParameters": {
                                        "summary": "InvalidParamters",
                                        "value": {
                                            "error": true,
                                            "message": "Invalid query parameters. Only country and populatedWithin are permitted."
                                        }
                                    },
                                    "InvalidPopulatedWithinParameter": {
                                        "summary": "InvalidPopulatedWithinParameter",
                                        "value": {
                                            "error": true,
                                            "message": "Invalid populatedWithin query parameter. Must be 5km, 10km, 30km, or 100km."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server / database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error in MySQL query"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/volcano/{id}": {
            "get": {
                "tags": [
                    "Data"
                ],
                "summary": "Retrieve data from a specific volcano",
                "description": "Returns an object containing data for the queried volcano. If a valid JWT token is sent in the header of the request, population data for the volcano is also provided. To test this using Swagger, click the green 'Authorize' button at the top of this page and enter your JWT token. A JWT token can be obtained by logging in. The path parameter (id) is required and represents the id of the volcano.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "description": "The ID of the volcano"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns an object containing the id, name, country, region, subregion, last eruption, summit, elevation, latitude, and longitude data for the queried volcano. If a valid JWT token is sent in the header of the request, the population data is also provided. An example of one object (with an authenticated request) is shown below.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Abu"
                                            },
                                            "country": {
                                                "type": "string",
                                                "example": "Japan"
                                            },
                                            "region": {
                                                "type": "string",
                                                "example": "Japan, Taiwan, Marianas"
                                            },
                                            "subregion": {
                                                "type": "string",
                                                "example": "Honshu"
                                            },
                                            "last_eruption": {
                                                "type": "string",
                                                "example": "6850 BCE"
                                            },
                                            "summit": {
                                                "type": "integer",
                                                "example": 641
                                            },
                                            "elevation": {
                                                "type": "integer",
                                                "example": 2103
                                            },
                                            "latitude": {
                                                "type": "string",
                                                "example": "34.5000"
                                            },
                                            "longitude": {
                                                "type": "string",
                                                "example": "131.6000"
                                            },
                                            "population_5km": {
                                                "type": "integer",
                                                "example": 3597
                                            },
                                            "population_10km": {
                                                "type": "integer",
                                                "example": 9594
                                            },
                                            "population_30km": {
                                                "type": "integer",
                                                "example": 117805
                                            },
                                            "population_100km": {
                                                "type": "integer",
                                                "example": 4071152
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server / database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error in MySQL query"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    }
}